+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_box_close_v2.py
@@ -30,8 +30,8 @@ class SawyerBoxCloseEnvV2(SawyerXYZEnv):
 
         self.init_config = {
             "obj_init_angle": 0.3,
-            "obj_init_pos": np.array([0, 0.55, 0.02], dtype=np.float32),
-            "hand_init_pos": np.array((0, 0.6, 0.2), dtype=np.float32),
+            "obj_init_pos": np.array([0, 0.55, 0.02], dtype=np.float64),
+            "hand_init_pos": np.array((0, 0.6, 0.2), dtype=np.float64),
         }
         self.goal = np.array([0.0, 0.75, 0.133])
         self.obj_init_pos = self.init_config["obj_init_pos"]
@@ -41,8 +41,8 @@ class SawyerBoxCloseEnvV2(SawyerXYZEnv):
         self._target_to_obj_init = None
 
         self.goal_space = Box(
-            np.array(goal_low, dtype=np.float32),
-            np.array(goal_high, dtype=np.float32),
+            np.array(goal_low, dtype=np.float64),
+            np.array(goal_high, dtype=np.float64),
         )
         self._random_reset_space = Box(
             np.hstack((obj_low, goal_low)),
