+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_hammer_v2.py
@@ -42,8 +42,8 @@ class SawyerHammerEnvV2(SawyerXYZEnv):
 
         self._random_reset_space = Box(np.array(obj_low), np.array(obj_high))
         self.goal_space = Box(
-            np.array(goal_low, dtype=np.float32),
-            np.array(goal_high, dtype=np.float32),
+            np.array(goal_low, dtype=np.float64),
+            np.array(goal_high, dtype=np.float64),
         )
 
     @property
