+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_stick_pull_v2.py
@@ -41,8 +41,8 @@ class SawyerStickPullEnvV2(SawyerXYZEnv):
         self.obj_init_qpos = np.array([0.0, 0.09])
         self.obj_space = Box(np.array(obj_low), np.array(obj_high))
         self.goal_space = Box(
-            np.array(goal_low, dtype=np.float32),
-            np.array(goal_high, dtype=np.float32),
+            np.array(goal_low, dtype=np.float64),
+            np.array(goal_high, dtype=np.float64),
         )
         self._random_reset_space = Box(
             np.hstack((obj_low, goal_low)),
